fix date props
