--- packages/react/src/auto/shadcn/types/calendarTypes.tsx
@@ -1,1340 +1,0 @@
-import type { CSSProperties } from "react";
-import { ComponentType } from "react";
-
-/**
- * The UI elements composing DayPicker. These elements are mapped to
- * {@link CustomComponents}, the {@link ClassNames} and the {@link Styles} used by
- * DayPicker.
- *
- * Some of these elements are extended by flags and modifiers.
- */
-export declare enum UI {
-  /** The root component displaying the months and the navigation bar. */
-  Root = "root",
-  /** The Chevron SVG element used by navigation buttons and dropdowns. */
-  Chevron = "chevron",
-  /**
-   * The grid cell with the day's date. Extended by {@link DayFlag} and
-   * {@link SelectionFlag}.
-   */
-  Day = "day",
-  /** The button containing the formatted day's date, inside the grid cell. */
-  DayButton = "day_button",
-  /** The caption label of the month (when not showing the dropdown navigation). */
-  CaptionLabel = "caption_label",
-  /** The container of the dropdown navigation (when enabled). */
-  Dropdowns = "dropdowns",
-  /** The dropdown element to select for years and months. */
-  Dropdown = "dropdown",
-  /** The container element of the dropdown. */
-  DropdownRoot = "dropdown_root",
-  /** The root element of the footer. */
-  Footer = "footer",
-  /** The month grid. */
-  MonthGrid = "month_grid",
-  /** Contains the dropdown navigation or the caption label. */
-  MonthCaption = "month_caption",
-  /** The dropdown with the months. */
-  MonthsDropdown = "months_dropdown",
-  /** Wrapper of the month grid. */
-  Month = "month",
-  /** The container of the displayed months. */
-  Months = "months",
-  /** The navigation bar with the previous and next buttons. */
-  Nav = "nav",
-  /**
-   * The next month button in the navigation. *
-   *
-   * @since 9.1.0
-   */
-  NextMonthButton = "button_next",
-  /**
-   * The previous month button in the navigation.
-   *
-   * @since 9.1.0
-   */
-  PreviousMonthButton = "button_previous",
-  /** The row containing the week. */
-  Week = "week",
-  /** The group of row weeks in a month (`tbody`). */
-  Weeks = "weeks",
-  /** The column header with the weekday. */
-  Weekday = "weekday",
-  /** The row grouping the weekdays in the column headers. */
-  Weekdays = "weekdays",
-  /** The cell containing the week number. */
-  WeekNumber = "week_number",
-  /** The cell header of the week numbers column. */
-  WeekNumberHeader = "week_number_header",
-  /** The dropdown with the years. */
-  YearsDropdown = "years_dropdown",
-}
-
-/** The flags for the {@link UI.Day}. */
-export declare enum DayFlag {
-  /** The day is disabled. */
-  disabled = "disabled",
-  /** The day is hidden. */
-  hidden = "hidden",
-  /** The day is outside the current month. */
-  outside = "outside",
-  /** The day is focused. */
-  focused = "focused",
-  /** The day is today. */
-  today = "today",
-}
-/**
- * The state that can be applied to the {@link UI.Day} element when in selection
- * mode.
- */
-export declare enum SelectionState {
-  /** The day is at the end of a selected range. */
-  range_end = "range_end",
-  /** The day is at the middle of a selected range. */
-  range_middle = "range_middle",
-  /** The day is at the start of a selected range. */
-  range_start = "range_start",
-  /** The day is selected. */
-  selected = "selected",
-}
-
-export type DayPickerProps = PropsBase &
-  (
-    | PropsSingle
-    | PropsSingleRequired
-    | PropsMulti
-    | PropsMultiRequired
-    | PropsRange
-    | PropsRangeRequired
-    | {
-        mode?: undefined;
-        required?: undefined;
-      }
-  );
-
-/**
- * Deprecated UI elements and flags.
- *
- * These elements were used in previous version of DayPicker and are kept here
- * to help the transition to the new {@link UI | UI elements}.
- *
- * ```diff
- *   <DayPicker classNames={{
- * -  cell: "my-cell",
- * +  day: "my-cell",
- * -  day: "my-day",
- * +  day_button: "my-day",
- * -  day_disabled: "my-day_disabled",
- * +  disabled: "my-day_disabled",
- *    // etc.
- *   }}/>
- * ```
- *
- * @deprecated
- * @since 9.0.1
- * @see https://daypicker.dev/upgrading
- * @see https://daypicker.dev/docs/styling
- */
-export type DeprecatedUI<T extends CSSProperties | string> = {
-  /**
-   * This element was applied to the style of any button in DayPicker and it is
-   * replaced by {@link UI.PreviousMonthButton} and {@link UI.NextMonthButton}.
-   *
-   * @deprecated
-   */
-  button: T;
-  /**
-   * This element was resetting the style of any button in DayPicker and it is
-   * replaced by {@link UI.PreviousMonthButton} and {@link UI.NextMonthButton}.
-   *
-   * @deprecated
-   */
-  button_reset: T;
-  /**
-   * This element has been renamed to {@link UI.MonthCaption}.
-   *
-   * @deprecated
-   */
-  caption: T;
-  /**
-   * This element has been removed. Captions are styled via
-   * {@link UI.MonthCaption}.
-   *
-   * @deprecated
-   */
-  caption_between: T;
-  /**
-   * This element has been renamed to {@link UI.Dropdowns}.
-   *
-   * @deprecated
-   */
-  caption_dropdowns: T;
-  /**
-   * This element has been removed. Captions are styled via
-   * {@link UI.MonthCaption}.
-   *
-   * @deprecated
-   */
-  caption_end: T;
-  /**
-   * This element has been removed.
-   *
-   * @deprecated
-   */
-  caption_start: T;
-  /**
-   * This element has been renamed to {@link UI.Day}.
-   *
-   * @deprecated
-   */
-  cell: T;
-  /**
-   * This element has been renamed to {@link DayFlag.disabled}.
-   *
-   * @deprecated
-   */
-  day_disabled: T;
-  /**
-   * This element has been renamed to {@link DayFlag.hidden}.
-   *
-   * @deprecated
-   */
-  day_hidden: T;
-  /**
-   * This element has been renamed to {@link DayFlag.outside}.
-   *
-   * @deprecated
-   */
-  day_outside: T;
-  /**
-   * This element has been renamed to {@link SelectionState.range_end}.
-   *
-   * @deprecated
-   */
-  day_range_end: T;
-  /**
-   * This element has been renamed to {@link SelectionState.range_middle}.
-   *
-   * @deprecated
-   */
-  day_range_middle: T;
-  /**
-   * This element has been renamed to {@link SelectionState.range_start}.
-   *
-   * @deprecated
-   */
-  day_range_start: T;
-  /**
-   * This element has been renamed to {@link SelectionState.selected}.
-   *
-   * @deprecated
-   */
-  day_selected: T;
-  /**
-   * This element has been renamed to {@link DayFlag.today}.
-   *
-   * @deprecated
-   */
-  day_today: T;
-  /**
-   * This element has been removed. The dropdown icon is now {@link UI.Chevron}
-   * inside a {@link UI.CaptionLabel}.
-   *
-   * @deprecated
-   */
-  dropdown_icon: T;
-  /**
-   * This element has been renamed to {@link UI.MonthsDropdown}.
-   *
-   * @deprecated
-   */
-  dropdown_month: T;
-  /**
-   * This element has been renamed to {@link UI.YearsDropdown}.
-   *
-   * @deprecated
-   */
-  dropdown_year: T;
-  /**
-   * This element has been removed.
-   *
-   * @deprecated
-   */
-  head: T;
-  /**
-   * This element has been renamed to {@link UI.Weekday}.
-   *
-   * @deprecated
-   */
-  head_cell: T;
-  /**
-   * This element has been renamed to {@link UI.Weekdays}.
-   *
-   * @deprecated
-   */
-  head_row: T;
-  /**
-   * This flag has been removed. Use `data-multiple-months` in your CSS
-   * selectors.
-   *
-   * @deprecated
-   */
-  multiple_months: T;
-  /**
-   * This element has been removed. To style the navigation buttons, use
-   * {@link UI.PreviousMonthButton} and {@link UI.NextMonthButton}.
-   *
-   * @deprecated
-   */
-  nav_button: T;
-  /**
-   * This element has been renamed to {@link UI.NextMonthButton}.
-   *
-   * @deprecated
-   */
-  nav_button_next: T;
-  /**
-   * This element has been renamed to {@link UI.PreviousMonthButton}.
-   *
-   * @deprecated
-   */
-  nav_button_previous: T;
-  /**
-   * This element has been removed. The dropdown icon is now {@link UI.Chevron}
-   * inside a {@link UI.NextMonthButton} or a {@link UI.PreviousMonthButton}.
-   *
-   * @deprecated
-   */
-  nav_icon: T;
-  /**
-   * This element has been renamed to {@link UI.Week}.
-   *
-   * @deprecated
-   */
-  row: T;
-  /**
-   * This element has been renamed to {@link UI.MonthGrid}.
-   *
-   * @deprecated
-   */
-  table: T;
-  /**
-   * This element has been renamed to {@link UI.Weeks}.
-   *
-   * @deprecated
-   */
-  tbody: T;
-  /**
-   * This element has been removed. The {@link UI.Footer} is now a single element
-   * below the months.
-   *
-   * @deprecated
-   */
-  tfoot: T;
-  /**
-   * This flag has been removed. There are no "visually hidden" elements in
-   * DayPicker 9.
-   *
-   * @deprecated
-   */
-  vhidden: T;
-  /**
-   * This element has been renamed. Use {@link UI.WeekNumber} instead.
-   *
-   * @deprecated
-   */
-  weeknumber: T;
-  /**
-   * This flag has been removed. Use `data-week-numbers` in your CSS.
-   *
-   * @deprecated
-   */
-  with_weeknumber: T;
-};
-
-/**
- * Selection modes supported by DayPicker.
- *
- * - `single`: use DayPicker to select single days.
- * - `multiple`: allow selecting multiple days.
- * - `range`: use DayPicker to select a range of days.
- *
- * @see https://daypicker.dev/docs/selection-modes
- */
-export type Mode = "single" | "multiple" | "range";
-/**
- * The components that can be changed using the `components` prop.
- *
- * @see https://daypicker.dev/guides/custom-components
- */
-export type CustomComponents = {
-  /**
-   * Render any button element in DayPicker.
-   *
-   * @deprecated Use {@link CustomComponents.NextMonthButton} or
-   *   {@link CustomComponents.PreviousMonthButton} instead.
-   */
-  Button: ComponentType;
-  /** Render the chevron icon used in the navigation buttons and dropdowns. */
-  Chevron: ComponentType;
-  /** Render the caption label of the month grid. */
-  CaptionLabel: ComponentType;
-  /** Render the day cell in the month grid. */
-  Day: ComponentType;
-  /** Render the button containing the day in the day cell. */
-  DayButton: ComponentType;
-  /** Render the dropdown element to select years and months. */
-  Dropdown: ComponentType;
-  /** Render the container of the dropdowns. */
-  DropdownNav: ComponentType;
-  /** Render the footer element announced by screen readers. */
-  Footer: ComponentType;
-  /** Render the container of the MonthGrid. */
-  Month: ComponentType;
-  /** Render the caption of the month grid. */
-  MonthCaption: ComponentType;
-  /** Render the grid of days in a month. */
-  MonthGrid: ComponentType;
-  /** Wrapper of the month grids. */
-  Months: ComponentType;
-  /** Render the navigation element with the next and previous buttons. */
-  Nav: ComponentType;
-  /** Render the `<option>` HTML element in the dropdown. */
-  Option: ComponentType;
-  /** Render the previous month button element in the navigation. */
-  PreviousMonthButton: ComponentType;
-  /** Render the next month button element in the navigation. */
-  NextMonthButton: ComponentType;
-  /** Render the root element of the calendar. */
-  Root: ComponentType;
-  /** Render the select element in the dropdowns. */
-  Select: ComponentType;
-  /** Render the weeks section in the month grid. */
-  Weeks: ComponentType;
-  /** Render the week rows. */
-  Week: ComponentType;
-  /** Render the weekday name in the header. */
-  Weekday: ComponentType;
-  /** Render the row containing the week days. */
-  Weekdays: ComponentType;
-  /** Render the cell with the number of the week. */
-  WeekNumber: ComponentType;
-  /** Render the header of the week number column. */
-  WeekNumberHeader: ComponentType;
-  /** Render the dropdown with the months. */
-  MonthsDropdown: ComponentType;
-  /** Render the dropdown with the years. */
-  YearsDropdown: ComponentType;
-};
-/** Represent a map of formatters used to render localized content. */
-export type Formatters = {
-  /** Format the caption of a month grid. */
-  formatCaption: (date: Date) => string;
-  /** Format the day in the day cell. */
-  formatDay: (date: Date) => string;
-  /** Format the label in the month dropdown. */
-  formatMonthDropdown: (date: Date) => string;
-  /**
-   * @ignore
-   * @deprecated Use {@link Formatters.formatCaption} instead.
-   */
-  formatMonthCaption: (date: Date) => string;
-  /** Format the week number. */
-  formatWeekNumber: (date: Date) => string;
-  /** Format the header of the week number column. */
-  formatWeekNumberHeader: (date: Date) => string;
-  /** Format the week day name in the header. */
-  formatWeekdayName: (date: Date) => string;
-  /** Format the label in the year dropdown. */
-  formatYearDropdown: (date: Date) => string;
-  /**
-   * @ignore
-   * @deprecated Use {@link Formatters.formatYearDropdown} instead.
-   */
-  formatYearCaption: (date: Date) => string;
-};
-/** Map of functions to translate ARIA labels for the relative elements. */
-export type Labels = {
-  /** The label for the navigation toolbar. */
-  labelNav: () => string;
-  /** The label for the month grid. */
-  labelGrid: () => string;
-  /** The label for the gridcell, when the calendar is not interactive. */
-  labelGridcell: () => string;
-  /** The label for the month dropdown. */
-  labelMonthDropdown: () => string;
-  /** The label for the year dropdown. */
-  labelYearDropdown: () => string;
-  /** The label for the "next month" button. */
-  labelNext: () => string;
-  /** The label for the "previous month" button. */
-  labelPrevious: () => string;
-  /** The label for the day button. */
-  labelDayButton: () => string;
-  /**
-   * @ignore
-   * @deprecated Use {@link labelDayButton} instead.
-   */
-  labelDay: () => string;
-  /** The label for the weekday. */
-  labelWeekday: () => string;
-  /** The label for the week number. */
-  labelWeekNumber: () => string;
-  /**
-   * Return the label for the column of the week number.
-   *
-   * @since 9.0.0
-   */
-  labelWeekNumberHeader: () => string;
-};
-/**
- * A value or a function that matches a specific day.
- *
- * @example
- *   // will always match the day
- *   const booleanMatcher: Matcher = true;
- *
- *   // will match the today's date
- *   const dateMatcher: Matcher = new Date();
- *
- *   // will match the days in the array
- *   const arrayMatcher: Matcher = [
- *     new Date(2019, 1, 2),
- *     new Date(2019, 1, 4)
- *   ];
- *
- *   // will match days after the 2nd of February 2019
- *   const afterMatcher: DateAfter = { after: new Date(2019, 1, 2) };
- *
- *   // will match days before the 2nd of February 2019 }
- *   const beforeMatcher: DateBefore = { before: new Date(2019, 1, 2) };
- *
- *   // will match Sundays
- *   const dayOfWeekMatcher: DayOfWeek = {
- *     dayOfWeek: 0
- *   };
- *
- *   // will match the included days, except the two dates
- *   const intervalMatcher: DateInterval = {
- *     after: new Date(2019, 1, 2),
- *     before: new Date(2019, 1, 5)
- *   };
- *
- *   // will match the included days, including the two dates
- *   const rangeMatcher: DateRange = {
- *     from: new Date(2019, 1, 2),
- *     to: new Date(2019, 1, 5)
- *   };
- *
- *   // will match when the function return true
- *   const functionMatcher: Matcher = (day: Date) => {
- *     return day.getMonth() === 2; // match when month is March
- *   };
- */
-export type Matcher = boolean | ((date: Date) => boolean) | Date | Date[] | DateRange | DateBefore | DateAfter | DateInterval | DayOfWeek;
-/**
- * Match a day falling after the specified date, with the date not included.
- *
- * @example
- *   // Match days after the 2nd of February 2019
- *   const matcher: DateAfter = { after: new Date(2019, 1, 2) };
- */
-export type DateAfter = {
-  after: Date;
-};
-/**
- * Match a day falling before the specified date, with the date not included.
- *
- * @example
- *   // Match days before the 2nd of February 2019
- *   const matcher: DateBefore = { before: new Date(2019, 1, 2) };
- */
-export type DateBefore = {
-  before: Date;
-};
-/**
- * An interval of dates. Differently from {@link DateRange}, the range ends here
- * are not included.
- *
- * @example
- *   // Match the days between the 2nd and the 5th of February 2019
- *   const matcher: DateInterval = {
- *     after: new Date(2019, 1, 2),
- *     before: new Date(2019, 1, 5)
- *   };
- */
-export type DateInterval = {
-  before: Date;
-  after: Date;
-};
-/**
- * A range of dates. The range can be open. Differently from
- * {@link DateInterval}, the range ends here are included.
- *
- * @example
- *   // Match the days between the 2nd and the 5th of February 2019
- *   const matcher: DateRange = {
- *     from: new Date(2019, 1, 2),
- *     to: new Date(2019, 1, 5)
- *   };
- */
-export type DateRange = {
-  from: Date | undefined;
-  to?: Date | undefined;
-};
-/**
- * Match dates being one of the specified days of the week (`0-6`, where `0` is
- * Sunday).
- *
- * @example
- *   // Match Sundays
- *   const matcher: DayOfWeek = { dayOfWeek: 0 };
- *   // Match weekends
- *   const matcher: DayOfWeek = { dayOfWeek: [0, 6] };
- */
-export type DayOfWeek = {
-  dayOfWeek: number | number[];
-};
-/**
- * The event handler triggered when clicking or interacting with a day.
- *
- * @template EventType - The event type that triggered the event (e.g.
- *   `React.MouseEvent`, `React.KeyboardEvent`, etc.).
- * @param date - The date that has triggered the event.
- * @param modifiers - The modifiers belonging to the date.
- * @param e - The DOM event that triggered the event.
- */
-export type DayEventHandler<EventType> = (date: Date, modifiers: Modifiers, e: EventType) => void;
-/**
- * The event handler when a month is changed in the calendar.
- *
- * ```tsx
- * <DayPicker onMonthChange={(month) => console.log(month)} />
- * ```
- *
- * @see https://daypicker.dev/docs/navigation
- */
-export type MonthChangeEventHandler = (month: Date) => void;
-/**
- * The CSS classnames to use for the {@link UI} elements, the
- * {@link SelectionState} and the {@link DayFlag}.
- *
- * @example
- *   const classNames: ClassNames = {
- *     [UI.Root]: "root",
- *     [UI.Outside]: "outside",
- *     [UI.Nav]: "nav"
- *     // etc.
- *   };
- */
-export type ClassNames = {
-  [key in UI | SelectionState | DayFlag]: string;
-};
-/**
- * The CSS styles to use for the {@link UI} elements, the {@link SelectionState}
- * and the {@link DayFlag}.
- */
-export type Styles = {
-  [key in UI | SelectionState | DayFlag]: CSSProperties | undefined;
-};
-/**
- * Represents the modifiers that match a specific day in the calendar.
- *
- * - Retrieve modifiers using the {@link OnSelectHandler} via the `onSelect` prop,
- *   or within custom components using the {@link useDayPicker} hook.
- * - Includes built-in modifiers from {@link DayFlag} and {@link SelectionState}.
- * - Add custom modifiers using the `modifiers` prop.
- *
- * @example
- *   const modifiers: Modifiers = {
- *   today: false, // the day is not today
- *   selected: true, // the day is selected
- *   disabled: false, // the day is not disabled
- *   outside: false, // the day is not outside the month
- *   focused: false, // the day is not focused
- *
- *   weekend: false // custom modifier example for matching a weekend
- *   booked: true // custom modifier example for matching a booked day
- *   available: false // custom modifier example for matching an available day
- *   };
- *
- * @see https://daypicker.dev/guides/custom-modifiers
- */
-export type Modifiers = Record<string, boolean>;
-/**
- * The style to apply to each day element matching a modifier.
- *
- * @example
- *   const modifiersStyles: ModifiersStyles = {
- *     today: { color: "red" },
- *     selected: { backgroundColor: "blue" },
- *     weekend: { color: "green" }
- *   };
- */
-export type ModifiersStyles = Record<string, CSSProperties>;
-/**
- * The classnames to assign to each day element matching a modifier.
- *
- * @example
- *   const modifiersClassNames: ModifiersClassNames = {
- *     today: "today", // Use the "today" class for the today's day
- *     selected: "highlight", // Use the "highlight" class for the selected day
- *     weekend: "weekend" // Use the "weekend" class for the weekend days
- *   };
- */
-export type ModifiersClassNames = Record<string, string>;
-/**
- * The props that have been deprecated since version 9.0.0.
- *
- * @private
- * @since 9.0.0
- * @see https://daypicker.dev/upgrading
- */
-export type V9DeprecatedProps =
-  /** Use `hidden` prop instead. */
-  | "fromDate"
-  /** Use `hidden` prop instead. */
-  | "toDate"
-  /** Use `startMonth` instead. */
-  | "fromMonth"
-  /** Use `endMonth` instead. */
-  | "toMonth"
-  /** Use `startMonth` instead. */
-  | "fromYear"
-  /** Use `endMonth` instead. */
-  | "toYear";
-/** The direction to move the focus relative to the current focused date. */
-export type MoveFocusDir = "after" | "before";
-/** The temporal unit to move the focus by. */
-export type MoveFocusBy = "day" | "week" | "startOfWeek" | "endOfWeek" | "month" | "year";
-/**
- * The numbering system supported by DayPicker.
- *
- * - `latn`: Latin (Western Arabic)
- * - `arab`: Arabic-Indic
- * - `arabext`: Eastern Arabic-Indic (Persian)
- * - `deva`: Devanagari
- * - `beng`: Bengali
- * - `guru`: Gurmukhi
- * - `gujr`: Gujarati
- * - `orya`: Oriya
- * - `tamldec`: Tamil
- * - `telu`: Telugu
- * - `knda`: Kannada
- * - `mlym`: Malayalam
- *
- * @see https://daypicker.dev/docs/translation#numeral-systems
- */
-export type Numerals = "latn" | "arab" | "arabext" | "deva" | "beng" | "guru" | "gujr" | "orya" | "tamldec" | "telu" | "knda" | "mlym";
-
-/**
- * Props for customizing the calendar, handling localization, and managing
- * events. These exclude the selection mode props.
- *
- * @group DayPicker
- * @see https://daypicker.dev/api/interfaces/PropsBase
- */
-export interface PropsBase {
-  /**
-   * Enable the selection of a single day, multiple days, or a range of days.
-   *
-   * @see https://daypicker.dev/docs/selection-modes
-   */
-  mode?: Mode | undefined;
-  /**
-   * Whether the selection is required.
-   *
-   * @see https://daypicker.dev/docs/selection-modes
-   */
-  required?: boolean | undefined;
-  /** Class name to add to the root element. */
-  className?: string;
-  /**
-   * Change the class names used by DayPicker.
-   *
-   * Use this prop when you need to change the default class names — for
-   * example, when importing the style via CSS modules or when using a CSS
-   * framework.
-   *
-   * @see https://daypicker.dev/docs/styling
-   */
-  classNames?: Partial<ClassNames> & Partial<DeprecatedUI<string>>;
-  /**
-   * Change the class name for the day matching the `modifiers`.
-   *
-   * @see https://daypicker.dev/guides/custom-modifiers
-   */
-  modifiersClassNames?: ModifiersClassNames;
-  /** Style to apply to the root element. */
-  style?: React.CSSProperties;
-  /**
-   * Change the inline styles of the HTML elements.
-   *
-   * @see https://daypicker.dev/docs/styling
-   */
-  styles?: Partial<Styles> & Partial<DeprecatedUI<React.CSSProperties>>;
-  /**
-   * Change the class name for the day matching the {@link modifiers}.
-   *
-   * @see https://daypicker.dev/guides/custom-modifiers
-   */
-  modifiersStyles?: ModifiersStyles;
-  /** A unique id to add to the root element. */
-  id?: string;
-  /**
-   * The initial month to show in the calendar.
-   *
-   * Use this prop to let DayPicker control the current month. If you need to
-   * set the month programmatically, use {@link month} and {@link onMonthChange}.
-   *
-   * @defaultValue The current month
-   * @see https://daypicker.dev/docs/navigation
-   */
-  defaultMonth?: Date;
-  /**
-   * The month displayed in the calendar.
-   *
-   * As opposed to `defaultMonth`, use this prop with `onMonthChange` to change
-   * the month programmatically.
-   *
-   * @see https://daypicker.dev/docs/navigation
-   */
-  month?: Date;
-  /**
-   * The number of displayed months.
-   *
-   * @defaultValue 1
-   * @see https://daypicker.dev/docs/customization#multiplemonths
-   */
-  numberOfMonths?: number;
-  /**
-   * The earliest month to start the month navigation.
-   *
-   * @since 9.0.0
-   * @see https://daypicker.dev/docs/navigation#start-and-end-dates
-   */
-  startMonth?: Date | undefined;
-  /**
-   * @private
-   * @deprecated This prop has been removed. Use `hidden={{ before: date }}`
-   *   instead.
-   * @see https://daypicker.dev/docs/navigation#start-and-end-dates
-   */
-  fromDate?: Date | undefined;
-  /**
-   * @private
-   * @deprecated This prop has been renamed to `startMonth`.
-   * @see https://daypicker.dev/docs/navigation#start-and-end-dates
-   */
-  fromMonth?: Date | undefined;
-  /**
-   * @private
-   * @deprecated Use `startMonth` instead. E.g. `startMonth={new Date(year,
-   *   0)}`.
-   * @see https://daypicker.dev/docs/navigation#start-and-end-dates
-   */
-  fromYear?: number | undefined;
-  /**
-   * The latest month to end the month navigation.
-   *
-   * @since 9.0.0
-   * @see https://daypicker.dev/docs/navigation#start-and-end-dates
-   */
-  endMonth?: Date;
-  /**
-   * @private
-   * @deprecated This prop has been removed. Use `hidden={{ after: date }}`
-   *   instead.
-   * @see https://daypicker.dev/docs/navigation#start-and-end-dates
-   */
-  toDate?: Date;
-  /**
-   * @private
-   * @deprecated This prop has been renamed to `endMonth`.
-   * @see https://daypicker.dev/docs/navigation#start-and-end-dates
-   */
-  toMonth?: Date;
-  /**
-   * @private
-   * @deprecated Use `endMonth` instead. E.g. `endMonth={new Date(year, 0)}`.
-   * @see https://daypicker.dev/docs/navigation#start-and-end-dates
-   */
-  toYear?: number;
-  /**
-   * Paginate the month navigation displaying the `numberOfMonths` at a time.
-   *
-   * @see https://daypicker.dev/docs/customization#multiplemonths
-   */
-  pagedNavigation?: boolean;
-  /**
-   * Render the months in reversed order (when {@link numberOfMonths} is set) to
-   * display the most recent month first.
-   *
-   * @see https://daypicker.dev/docs/customization#multiplemonths
-   */
-  reverseMonths?: boolean;
-  /**
-   * Hide the navigation buttons. This prop won't disable the navigation: to
-   * disable the navigation, use {@link disableNavigation}.
-   *
-   * @since 9.0.0
-   * @see https://daypicker.dev/docs/navigation#hidenavigation
-   */
-  hideNavigation?: boolean;
-  /**
-   * Disable the navigation between months. This prop won't hide the navigation:
-   * to hide the navigation, use {@link hideNavigation}.
-   *
-   * @see https://daypicker.dev/docs/navigation#disablenavigation
-   */
-  disableNavigation?: boolean;
-  /**
-   * Show dropdowns to navigate between months or years.
-   *
-   * - `true`: display the dropdowns for both month and year
-   * - `label`: display the month and the year as a label. Change the label with
-   *   the `formatCaption` formatter.
-   * - `month`: display only the dropdown for the months
-   * - `year`: display only the dropdown for the years
-   *
-   * **Note:** showing the dropdown will set the start/end months
-   * {@link startMonth} to 100 years ago, and {@link endMonth} to the end of the
-   * current year.
-   *
-   * @see https://daypicker.dev/docs/customization#caption-layouts
-   */
-  captionLayout?: "label" | "dropdown" | "dropdown-months" | "dropdown-years";
-  /**
-   * Display always 6 weeks per each month, regardless of the month’s number of
-   * weeks. Weeks will be filled with the days from the next month.
-   *
-   * @see https://daypicker.dev/docs/customization#fixed-weeks
-   */
-  fixedWeeks?: boolean;
-  /**
-   * Hide the row displaying the weekday row header.
-   *
-   * @since 9.0.0
-   */
-  hideWeekdays?: boolean;
-  /**
-   * Show the outside days (days falling in the next or the previous month).
-   *
-   * **Note:** when a broadcast {@link calendar} is set, this prop defaults to
-   * true.
-   *
-   * @see https://daypicker.dev/docs/customization#outside-days
-   */
-  showOutsideDays?: boolean;
-  /**
-   * Show the week numbers column. Weeks are numbered according to the local
-   * week index.
-   *
-   * @see https://daypicker.dev/docs/customization#showweeknumber
-   */
-  showWeekNumber?: boolean;
-  /**
-   * Display the weeks in the month following the broadcast calendar. Setting
-   * this prop will ignore {@link weekStartsOn} (always Monday) and
-   * {@link showOutsideDays} will default to true.
-   *
-   * @since 9.4.0
-   * @see https://daypicker.dev/docs/localization#broadcast-calendar
-   * @see https://en.wikipedia.org/wiki/Broadcast_calendar
-   */
-  broadcastCalendar?: boolean;
-  /**
-   * Use ISO week dates instead of the locale setting. Setting this prop will
-   * ignore `weekStartsOn` and `firstWeekContainsDate`.
-   *
-   * @see https://daypicker.dev/docs/localization#iso-week-dates
-   * @see https://en.wikipedia.org/wiki/ISO_week_date
-   */
-  ISOWeek?: boolean;
-  /**
-   * The time zone (IANA or UTC offset) to use in the calendar (experimental).
-   * See
-   * [Wikipedia](https://en.wikipedia.org/wiki/List_of_tz_database_time_zones)
-   * for the possible values.
-   *
-   * Time zones are supported by the `TZDate` object by the
-   * [@date-fns/tz](https://github.com/date-fns/tz) package. Please refer to the
-   * package documentation for more information.
-   *
-   * @since 9.1.1
-   * @experimental
-   * @see https://daypicker.dev/docs/time-zone
-   */
-  timeZone?: string | undefined;
-  /**
-   * Change the components used for rendering the calendar elements.
-   *
-   * @see https://daypicker.dev/guides/custom-components
-   */
-  components?: Partial<CustomComponents>;
-  /**
-   * Add a footer to the calendar, acting as a live region.
-   *
-   * Use this prop to communicate the calendar's status to screen readers.
-   * Prefer strings over complex UI elements.
-   *
-   * @see https://daypicker.dev/guides/accessibility#footer
-   */
-  footer?: React.ReactNode | string;
-  /**
-   * When a selection mode is set, DayPicker will focus the first selected day
-   * (if set) or today's date (if not disabled).
-   *
-   * Use this prop when you need to focus DayPicker after a user action, for
-   * improved accessibility.
-   *
-   * @see https://daypicker.dev/guides/accessibility#autofocus
-   */
-  autoFocus?: boolean;
-  /**
-   * @private
-   * @deprecated This prop will be removed. Use {@link autoFocus} instead.
-   */
-  initialFocus?: boolean;
-  /**
-   * Apply the `disabled` modifier to the matching days.
-   *
-   * @see https://daypicker.dev/docs/selection-modes#disabling-dates
-   */
-  disabled?: Matcher | Matcher[] | undefined;
-  /**
-   * Apply the `hidden` modifier to the matching days. Will hide them from the
-   * calendar.
-   *
-   * @see https://daypicker.dev/guides/custom-modifiers#hidden-modifier
-   */
-  hidden?: Matcher | Matcher[] | undefined;
-  /**
-   * The today’s date. Default is the current date. This date will get the
-   * `today` modifier to style the day.
-   *
-   * @see https://daypicker.dev/guides/custom-modifiers#today-modifier
-   */
-  today?: Date;
-  /**
-   * Add modifiers to the matching days.
-   *
-   * @see https://daypicker.dev/guides/custom-modifiers
-   */
-  modifiers?: Record<string, Matcher | Matcher[] | undefined> | undefined;
-  /**
-   * Labels creators to override the defaults. Use this prop to customize the
-   * aria-label attributes in DayPicker.
-   *
-   * @see https://daypicker.dev/docs/translation#aria-labels
-   */
-  labels?: Partial<Labels>;
-  /**
-   * Formatters used to format dates to strings. Use this prop to override the
-   * default functions.
-   *
-   * @see https://daypicker.dev/docs/translation#custom-formatters
-   */
-  formatters?: Partial<Formatters>;
-  /**
-   * The text direction of the calendar. Use `ltr` for left-to-right (default)
-   * or `rtl` for right-to-left.
-   *
-   * @see https://daypicker.dev/docs/translation#rtl-text-direction
-   */
-  dir?: HTMLDivElement["dir"];
-  /**
-   * The aria-label attribute to add to the container element.
-   *
-   * @since 9.4.1
-   * @see https://daypicker.dev/guides/accessibility
-   */
-  ["aria-label"]?: string;
-  /**
-   * The role attribute to add to the container element.
-   *
-   * @since 9.4.1
-   * @see https://daypicker.dev/guides/accessibility
-   */
-  role?: "application" | "dialog" | undefined;
-  /**
-   * A cryptographic nonce ("number used once") which can be used by Content
-   * Security Policy for the inline `style` attributes.
-   */
-  nonce?: HTMLDivElement["nonce"];
-  /** Add a `title` attribute to the container element. */
-  title?: HTMLDivElement["title"];
-  /** Add the language tag to the container element. */
-  lang?: HTMLDivElement["lang"];
-  /**
-   * The locale object used to localize dates. Pass a locale from
-   * `react-day-picker/locale` to localize the calendar.
-   *
-   * @example
-   *   import { es } from "react-day-picker/locale";
-   *   <DayPicker locale={es} />
-   *
-   * @defaultValue enUS - The English locale default of `date-fns`.
-   * @see https://daypicker.dev/docs/localization
-   * @see https://github.com/date-fns/date-fns/tree/main/src/locale for a list of the supported locales
-   */
-  locale?: Partial<Locale> | undefined;
-  /**
-   * The numeral system to use when formatting dates.
-   *
-   * - `latn`: Latin (Western Arabic)
-   * - `arab`: Arabic-Indic
-   * - `arabext`: Eastern Arabic-Indic (Persian)
-   * - `deva`: Devanagari
-   * - `beng`: Bengali
-   * - `guru`: Gurmukhi
-   * - `gujr`: Gujarati
-   * - `orya`: Oriya
-   * - `tamldec`: Tamil
-   * - `telu`: Telugu
-   * - `knda`: Kannada
-   * - `mlym`: Malayalam
-   *
-   * @defaultValue `latn` Latin (Western Arabic)
-   * @see https://daypicker.dev/docs/translation#numeral-systems
-   */
-  numerals?: Numerals | undefined;
-  /**
-   * The index of the first day of the week (0 - Sunday). Overrides the locale's
-   * one.
-   *
-   * @see https://daypicker.dev/docs/localization#first-date-of-the-week
-   */
-  weekStartsOn?: 0 | 1 | 2 | 3 | 4 | 5 | 6 | undefined;
-  /**
-   * The day of January, which is always in the first week of the year.
-   *
-   * @see https://daypicker.dev/docs/localization#first-week-contains-date
-   */
-  firstWeekContainsDate?: 1 | 4;
-  /**
-   * Enable `DD` and `DDDD` for week year tokens when formatting or parsing
-   * dates.
-   *
-   * @see https://date-fns.org/docs/Unicode-Tokens
-   */
-  useAdditionalWeekYearTokens?: boolean | undefined;
-  /**
-   * Enable `YY` and `YYYY` for day of year tokens when formatting or parsing
-   * dates.
-   *
-   * @see https://date-fns.org/docs/Unicode-Tokens
-   */
-  useAdditionalDayOfYearTokens?: boolean | undefined;
-  /**
-   * Event fired when the user navigates between months.
-   *
-   * @see https://daypicker.dev/docs/navigation#onmonthchange
-   */
-  onMonthChange?: MonthChangeEventHandler;
-  /**
-   * Event handler when the next month button is clicked.
-   *
-   * @see https://daypicker.dev/docs/navigation
-   */
-  onNextClick?: MonthChangeEventHandler;
-  /**
-   * Event handler when the previous month button is clicked.
-   *
-   * @see https://daypicker.dev/docs/navigation
-   */
-  onPrevClick?: MonthChangeEventHandler;
-  /**
-   * Event handler when a week number is clicked.
-   *
-   * @private
-   * @deprecated Use a custom `WeekNumber` component instead.
-   * @see https://daypicker.dev/docs/customization#showweeknumber
-   */
-  onWeekNumberClick?: any;
-  /** Event handler when a day is clicked. */
-  onDayClick?: DayEventHandler<React.MouseEvent>;
-  /** Event handler when a day is focused. */
-  onDayFocus?: DayEventHandler<React.FocusEvent>;
-  /** Event handler when a day is blurred. */
-  onDayBlur?: DayEventHandler<React.FocusEvent>;
-  /** Event handler when a key is pressed on a day. */
-  onDayKeyDown?: DayEventHandler<React.KeyboardEvent>;
-  /** Event handler when the mouse enters a day. */
-  onDayMouseEnter?: DayEventHandler<React.MouseEvent>;
-  /** Event handler when the mouse leaves a day. */
-  onDayMouseLeave?: DayEventHandler<React.MouseEvent>;
-  /**
-   * Replace the default date library with a custom one. Experimental: not
-   * guaranteed to be stable (may not respect semver).
-   *
-   * @since 9.0.0
-   * @experimental
-   */
-  dateLib?: any;
-  /**
-   * @private
-   * @deprecated Use a custom `DayButton` component instead.
-   */
-  onDayKeyUp?: DayEventHandler<React.KeyboardEvent>;
-  /**
-   * @private
-   * @deprecated Use a custom `DayButton` component instead.
-   */
-  onDayKeyPress?: DayEventHandler<React.KeyboardEvent>;
-  /**
-   * @private
-   * @deprecated Use a custom `DayButton` component instead.
-   */
-  onDayPointerEnter?: DayEventHandler<React.PointerEvent>;
-  /**
-   * @private
-   * @deprecated Use a custom `DayButton` component instead.
-   */
-  onDayPointerLeave?: DayEventHandler<React.PointerEvent>;
-  /**
-   * @private
-   * @deprecated Use a custom `DayButton` component instead.
-   */
-  onDayTouchCancel?: DayEventHandler<React.TouchEvent>;
-  /**
-   * @private
-   * @deprecated Use a custom `DayButton` component instead.
-   */
-  onDayTouchEnd?: DayEventHandler<React.TouchEvent>;
-  /**
-   * @private
-   * @deprecated Use a custom `DayButton` component instead.
-   */
-  onDayTouchMove?: DayEventHandler<React.TouchEvent>;
-  /**
-   * @private
-   * @deprecated Use a custom `DayButton` component instead.
-   */
-  onDayTouchStart?: DayEventHandler<React.TouchEvent>;
-}
-/**
- * Shared handler type for `onSelect` callback when a selection mode is set.
- *
- * @template T - The type of the selected item.
- * @callback OnSelectHandler
- * @param {T} selected - The selected item after the event.
- * @param {Date} triggerDate - The date when the event was triggered.
- * @param {Modifiers} modifiers - The modifiers associated with the event.
- * @param {React.MouseEvent | React.KeyboardEvent} e - The event object.
- */
-export type OnSelectHandler<T> = (selected: T, triggerDate: Date, modifiers: Modifiers, e: React.MouseEvent | React.KeyboardEvent) => void;
-/**
- * The props when the single selection is required.
- *
- * @group DayPicker
- * @see https://daypicker.dev/docs/selection-modes#single-mode
- */
-export interface PropsSingleRequired {
-  mode: "single";
-  required: true;
-  /** The selected date. */
-  selected: Date | undefined;
-  /** Event handler when a day is selected. */
-  onSelect?: OnSelectHandler<Date>;
-}
-/**
- * The props when the single selection is optional.
- *
- * @group DayPicker
- * @see https://daypicker.dev/docs/selection-modes#single-mode
- */
-export interface PropsSingle {
-  mode: "single";
-  required?: false | undefined;
-  /** The selected date. */
-  selected?: Date | undefined;
-  /** Event handler when a day is selected. */
-  onSelect?: OnSelectHandler<Date | undefined>;
-}
-/**
- * The props when the multiple selection is required.
- *
- * @group DayPicker
- * @see https://daypicker.dev/docs/selection-modes#multiple-mode
- */
-export interface PropsMultiRequired {
-  mode: "multiple";
-  required: true;
-  /** The selected dates. */
-  selected: Date[] | undefined;
-  /** Event handler when days are selected. */
-  onSelect?: OnSelectHandler<Date[]>;
-  /** The minimum number of selectable days. */
-  min?: number;
-  /** The maximum number of selectable days. */
-  max?: number;
-}
-/**
- * The props when the multiple selection is optional.
- *
- * @group DayPicker
- * @see https://daypicker.dev/docs/selection-modes#multiple-mode
- */
-export interface PropsMulti {
-  mode: "multiple";
-  required?: false | undefined;
-  /** The selected dates. */
-  selected?: Date[] | undefined;
-  /** Event handler when days are selected. */
-  onSelect?: OnSelectHandler<Date[] | undefined>;
-  /** The minimum number of selectable days. */
-  min?: number;
-  /** The maximum number of selectable days. */
-  max?: number;
-}
-/**
- * The props when the range selection is required.
- *
- * @group DayPicker
- * @see https://daypicker.dev/docs/selection-modes#range-mode
- */
-export interface PropsRangeRequired {
-  mode: "range";
-  required: true;
-  disabled?: Matcher | Matcher[] | undefined;
-  /**
-   * When `true`, the range will reset when including a disabled day.
-   *
-   * @since V9.0.2
-   */
-  excludeDisabled?: boolean | undefined;
-  /** The selected range. */
-  selected: DateRange | undefined;
-  /** Event handler when a range is selected. */
-  onSelect?: OnSelectHandler<DateRange>;
-  /** The minimum number of days to include in the range. */
-  min?: number;
-  /** The maximum number of days to include in the range. */
-  max?: number;
-}
-/**
- * The props when the range selection is optional.
- *
- * @group DayPicker
- * @see https://daypicker.dev/docs/selection-modes#range-mode
- */
-export interface PropsRange {
-  mode: "range";
-  required?: false | undefined;
-  disabled?: Matcher | Matcher[] | undefined;
-  /**
-   * When `true`, the range will reset when including a disabled day.
-   *
-   * @since V9.0.2
-   */
-  excludeDisabled?: boolean | undefined;
-  /** The selected range. */
-  selected?: DateRange | undefined;
-  /** Event handler when the selection changes. */
-  onSelect?: OnSelectHandler<DateRange | undefined>;
-  /** The minimum number of days to include in the range. */
-  min?: number;
-  /** The maximum number of days to include in the range. */
-  max?: number;
-}
